Implements a process-safe way to load wx.
