+++ MAVProxy/modules/lib/mp_checklist.py
@@ -5,8 +5,9 @@
   Created by Stephen Dade (stephen_dade@hotmail.com)
 """
 
-import wx, sys
+import sys
 from MAVProxy.modules.lib import mp_util
+from wx_loader import wx
 
 class CheckItem():
     '''Checklist item used for information transfer
@@ -33,9 +34,9 @@ class CheckUI():
     def child_task(self):
         '''child process - this holds all the GUI elements'''
         mp_util.child_close_fds()
-        import wx
+        from wx_loader import wx
         
-        app = wx.PySimpleApp()
+        app = wx.App(False)
         app.frame = ChecklistFrame(state=self, title=self.title)
         app.frame.Show()
         app.MainLoop()
@@ -364,5 +365,3 @@ if __name__ == "__main__":
     while checklist.is_alive():
         checklist.set_check("Compass Calibrated", 1)
         time.sleep(0.5)
-        
-
