+++ MAVProxy/modules/lib/mp_image.py
@@ -6,7 +6,7 @@ June 2012
 '''
 
 import time
-import wx
+from wx_loader import wx
 
 try:
     import cv2.cv as cv
@@ -104,10 +104,10 @@ class MPImage():
     def child_task(self):
         '''child process - this holds all the GUI elements'''
         mp_util.child_close_fds()
-        import wx
+        from wx_loader import wx
         state = self
 
-        self.app = wx.PySimpleApp()
+        self.app = wx.App(False)
         self.app.frame = MPImageFrame(state=self)
         self.app.frame.Show()
         self.app.MainLoop()
