+++ MAVProxy/modules/lib/optparse_gui/__init__.py
@@ -9,8 +9,7 @@ import os
 import sys
 import re
 import optparse
-
-import wx
+from wx_loader import wx
 
 __version__ = 0.1
 __revision__ = '$Id$'
