+++ MAVProxy/modules/lib/wxconsole_util.py
@@ -1,0 +1,15 @@
+class Text():
+    '''text to write to console'''
+    def __init__(self, text, fg='black', bg='white'):
+        self.text = text
+        self.fg = fg
+        self.bg = bg
+
+class Value():
+    '''a value for the status bar'''
+    def __init__(self, name, text, row=0, fg='black', bg='white'):
+        self.name = name
+        self.text = text
+        self.row = row
+        self.fg = fg
+        self.bg = bg
