+++ MAVProxy/modules/mavproxy_misseditor/missionEditorFrame.py
@@ -4,8 +4,7 @@
 # generated by wxGlade 0.6.8 on Wed Jun 11 13:41:49 2014
 #
 
-import wx
-import wx.grid
+from wx_loader.wx import grid
 
 # begin wxGlade: dependencies
 # end wxGlade
@@ -646,7 +645,7 @@ class MissionEditorFrame(wx.Frame):
         event.Skip()
 # end of class MissionEditorFrame
 if __name__ == "__main__":
-    app = wx.PySimpleApp(0)
+    app = wx.App(False)
     wx.InitAllImageHandlers()
     missionEditor = (None, wx.ID_ANY, "")
     app.SetTopWindow(missionEditor)
