+++ MAVProxy/modules/mavproxy_rally.py
@@ -30,11 +30,11 @@ class RallyModule(mp_module.MPModule):
                                   items=[MPMenuItem('Clear', 'Clear', '# rally clear'),
                                          MPMenuItem('List', 'List', '# rally list'),
                                          MPMenuItem('Load', 'Load', '# rally load ',
-                                                    handler=MPMenuCallFileDialog(flags=wx.FD_OPEN,
+                                                    handler=MPMenuCallFileDialog(flags=('open',),
                                                                                  title='Rally Load',
                                                                                  wildcard='*.rally')),
                                          MPMenuItem('Save', 'Save', '# rally save ',
-                                                    handler=MPMenuCallFileDialog(flags=wx.FD_SAVE|wx.FD_OVERWRITE_PROMPT,
+                                                    handler=MPMenuCallFileDialog(flags=('save', 'overwrite_prompt'),
                                                                                  title='Rally Save',
                                                                                  wildcard='*.rally')),
                                          MPMenuItem('Add', 'Add', '# rally add ',
