+++ MAVProxy/modules/mavproxy_wp.py
@@ -38,11 +38,11 @@ class WPModule(mp_module.MPModule):
                                   items=[MPMenuItem('Clear', 'Clear', '# wp clear'),
                                          MPMenuItem('List', 'List', '# wp list'),
                                          MPMenuItem('Load', 'Load', '# wp load ',
-                                                    handler=MPMenuCallFileDialog(flags=wx.FD_OPEN,
+                                                    handler=MPMenuCallFileDialog(flags=('open',),
                                                                                  title='Mission Load',
                                                                                  wildcard='*.txt')),
                                          MPMenuItem('Save', 'Save', '# wp save ',
-                                                    handler=MPMenuCallFileDialog(flags=wx.FD_SAVE|wx.FD_OVERWRITE_PROMPT,
+                                                    handler=MPMenuCallFileDialog(flags=('save', 'overwrite_prompt'),
                                                                                  title='Mission Save',
                                                                                  wildcard='*.txt')),
                                          MPMenuItem('Draw', 'Draw', '# wp draw ',
